+++ src/main/kotlin/com/github/mraghurao92/javadocgenerator/model/Message.kt
@@ -1,4 +1,4 @@
-package com.github.mraghurao92.javadocgenerator.javadoc
+package com.github.mraghurao92.javadocgenerator.model
 
 class Message {
     var role: String? = null