+++ CHANGELOG.md
@@ -1,7 +1,16 @@
-<!-- Keep a Changelog guide -> https://keepachangelog.com -->
+# Changelog
 
-# javadoc-generator Changelog
+All notable changes to the JavaDoc Generator IntelliJ Plugin will be documented in this file.
 
 ## [Unreleased]
+
 ### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+
+- JavaDoc Generation: This plugin utilizes the capabilities of ChatGPT AI to automatically generate JavaDoc documentation for a provided input method.
+
+## [0.0.1] - Not Yet Published
+
+### Features
+
+- JavaDoc Generation: This plugin leverages the capabilities of ChatGPT AI to produce JavaDoc documentation for a provided input method.
+